refactor: TripEvent 분리 Side Effect 처리[AIR-136]
