+++ aircnc/src/test/java/com/gurudev/aircnc/domain/trip/service/TripServiceForSchedulerImplTest.java
@@ -4,7 +4,7 @@ import static com.gurudev.aircnc.domain.trip.entity.TripStatus.RESERVED;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.gurudev.aircnc.domain.trip.entity.Trip;
-import com.gurudev.aircnc.domain.trip.service.command.TripCommand.TripEvent;
+import com.gurudev.aircnc.infrastructure.event.TripEvent;
 import java.util.List;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,8 +17,8 @@ class TripServiceForSchedulerImplTest extends BaseTripServiceTest {
   @Test
   void 여행_상태_bulk_업데이트_테스트() {
     // given
-    TripEvent tripEvent1 = defaultTripReserveCommand();
-    TripEvent tripEvent2 = defaultTripReserveCommand();
+    TripEvent tripEvent1 = defaultTripEvent();
+    TripEvent tripEvent2 = defaultTripEvent();
 
     tripService.reserve(tripEvent1);
     tripService.reserve(tripEvent2);
